新增异步方法及更新删除逻辑
在 `IPointsInfoRepository.cs` 文件中，添加了两个新的异步方法 `ResetPointsAsync` 和 `ResetSignInAsync`。
在 `PointsInfoRepository.cs` 文件中，实现了 `ResetPointsAsync` 和 `ResetSignInAsync` 方法，分别用于将 `Points` 字段重置为0和将 `LastSignInAt` 字段重置为NULL。
在 `ChatRecordController.cs` 文件中，`Delete` 方法增加了一个 `deleteAll` 参数，并根据该参数决定是删除所有记录还是根据 `ids` 删除指定记录。
在 `HomeLocationController.cs` 文件中，`Delete` 方法增加了一个 `deleteAll` 参数，并根据该参数决定是删除所有记录还是根据 `ids` 删除指定记录。
在 `PointsInfoController.cs` 文件中，`Delete` 方法增加了 `resetSignIn` 和 `resetPoints` 参数，并根据这些参数决定是重置签到时间、重置积分还是根据 `ids` 删除指定记录。
在 `PointsInfoController.cs` 文件的 `using` 部分，添加了对 `SdtdServerKit.Data.Repositories` 命名空间的引用。

+++ src/SdtdServerKit/WebApi/Controllers/PointsInfoController.cs
@@ -1,6 +1,7 @@
 ﻿using IceCoffee.SimpleCRUD.Dtos;
 using SdtdServerKit.Data.Entities;
 using SdtdServerKit.Data.IRepositories;
+using SdtdServerKit.Data.Repositories;
 using System.ComponentModel.DataAnnotations;
 
 namespace SdtdServerKit.WebApi.Controllers
@@ -121,12 +122,29 @@ namespace SdtdServerKit.WebApi.Controllers
         /// 批量删除记录
         /// </summary>
         /// <param name="ids"></param>
+        /// <param name="resetSignIn"></param>
+        /// <param name="resetPoints"></param>
         /// <returns></returns>
         [HttpDelete]
         [Route("")]
-        public async Task<IHttpActionResult> Delete([MinLength(1)] string[] ids)
+        public async Task<IHttpActionResult> Delete([FromUri] string[]? ids, [FromUri] bool resetSignIn = false, [FromUri] bool resetPoints = false)
         {
-            int count = await _repository.DeleteByIdsAsync(ids, true);
+            int count = 0;
+
+            if (resetSignIn)
+            {
+                count = await _repository.ResetSignInAsync();
+            }
+            else if (resetPoints)
+            {
+                count = await _repository.ResetPointsAsync();
+            }
+            else if (ids != null && ids.Length > 0)
+            {
+                count = await _repository.DeleteByIdsAsync(ids, true);
+            }
+
+
             return Ok(count);
         }
     }
